Backport hotfix-1.6.x changes
