+++ databases/orientdb/src/main/java/com/gentics/mesh/graphdb/tx/impl/OrientServerStorageImpl.java
@@ -7,12 +7,10 @@ import org.apache.commons.lang3.NotImplementedException;
 import com.gentics.mesh.etc.config.MeshOptions;
 import com.gentics.mesh.graphdb.tx.AbstractOrientStorage;
 import com.gentics.mesh.metric.MetricsService;
-import com.orientechnologies.orient.core.Orient;
 import com.orientechnologies.orient.core.db.ODatabaseDocumentInternal;
 import com.orientechnologies.orient.core.db.ODatabaseSession;
 import com.orientechnologies.orient.core.db.ODatabaseType;
 import com.orientechnologies.orient.core.db.OrientDB;
-import com.orientechnologies.orient.core.db.OrientDBConfig;
 import com.tinkerpop.blueprints.impls.orient.OrientGraph;
 import com.tinkerpop.blueprints.impls.orient.OrientGraphFactory;
 import com.tinkerpop.blueprints.impls.orient.OrientGraphNoTx;
@@ -36,7 +34,6 @@ public class OrientServerStorageImpl extends AbstractOrientStorage {
 
 	@Override
 	public void close() {
-		Orient.instance().shutdown();
 	}
 
 	@Override
